requirements.txt
streamlit 
pyyaml 
google-generativeai
matplotlib